atomisp2: CSS1.5 20121027_0807 release for MRFLD
BZ: 66183

This change is the incremental patch for MRFLD CSS1.5.
20121003_1834 -> 20121027_0807

Reviewed-on: http://android.intel.com:8080/72594
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: lab_cactus <lab_cactus@intel.com>

+++ drivers/media/video/atomisp2/css2400/sh_css_rx.c
@@ -265,7 +265,7 @@ enum sh_css_err sh_css_input_format_type(
 		break;
 #if defined(HAS_RX_VERSION_2)
 	default:
-		if (input_format > N_MIPI_FORMAT)
+		if (input_format > (unsigned int)N_MIPI_FORMAT)
 			return sh_css_err_internal_error;
 		*fmt_type = input_format;
 		break;
